telegram_bot/requirements.txt
python-telegram-bot==22.6
httpx
python-dotenv
requests